nn working, ~ 80%

+++ trySKall.py
@@ -23,48 +23,66 @@ if __name__ == '__main__':
     
 
     classifiers = [
-        KNeighborsClassifier(3),
-        SVC(kernel="linear", C=0.025),
-        SVC(gamma=2, C=1),
+#        KNeighborsClassifier(10),
+        SVC(kernel="linear", C=0.05),
+        SVC(gamma="scale", C=0.05),
 #        GaussianProcessClassifier(1.0 * RBF(1.0)),
-        DecisionTreeClassifier(max_depth=5),
-        RandomForestClassifier(max_depth=5, n_estimators=10, max_features=1),
+#        DecisionTreeClassifier(max_depth=10),
+#        RandomForestClassifier(max_depth=10, n_estimators=10),
         MLPClassifier(alpha=1),
-        AdaBoostClassifier(),
-        GaussianNB(),
-        QuadraticDiscriminantAnalysis()]
+        # AdaBoostClassifier(),
+        # GaussianNB(),
+        #QuadraticDiscriminantAnalysis()
+        ]
 
-    names = ["Nearest Neighbors", "Linear SVM", "RBF SVM",# "Gaussian Process",
-            "Decision Tree", "Random Forest", "Neural Net", "AdaBoost",
-            "Naive Bayes", "QDA"]
+    names = [#"Nearest Neighbors", 
+             "Linear SVM", 
+             "RBF SVM",
+             # "Gaussian Process",
+ #           "Decision Tree", 
+ #           "Random Forest", 
+            "Neural Net", 
+            #"AdaBoost",
+            #"Naive Bayes", 
+            #"QDA"
+            ]
 
     XTrain=[]
     YTrain=[]
     XTest=[]
     YTest=[]
 
+
     panels=sorted(glob(join(sys.argv[1]+'*')))
     p2i={p:i for i,p in enumerate([basename(x) for x in panels])}
     i2p={p2i[p] for p in p2i.keys()}
-    for p in panels:
-        cf=glob(join(p,'t_*.fv'))
-        b=[basename(x) for x in cf]
-        prefs=list(set([x[x.find('_'):x.rfind('_')+1] for x in b]))
-        testnames=sample(prefs,int(np.ceil(0.2*len(prefs))))
-        for f in cf:
-            if any([(x in f) for x in testnames]):
-                XTest.append(np.loadtxt(f))
-                YTest.append(p2i[basename(p)])
-            else:
-                XTrain.append(np.loadtxt(f))
-                YTrain.append(p2i[basename(p)])
 
-    XTest=np.array(XTest)
-    YTest=np.array(YTest)
-    XTrain=np.array(XTrain)
-    YTrain=np.array(YTrain)
+    if exists('train.npz'):
+        data=np.load('train.npz')
+        XTest=data['XTest']
+        YTest=data['YTest']
+        XTrain=data['XTrain']
+        YTrain=data['YTrain']
+    else:
+        for p in panels:
+            cf=glob(join(p,'t_*.fv'))
+            b=[basename(x) for x in cf]
+            prefs=list(set([x[x.find('_'):x.rfind('_')+1] for x in b]))
+            testnames=sample(prefs,int(np.ceil(0.2*len(prefs))))
+            for f in cf:
+                if any([(x in f) for x in testnames]):
+                    XTest.append(np.loadtxt(f))
+                    YTest.append(p2i[basename(p)])
+                else:
+                    XTrain.append(np.loadtxt(f))
+                    YTrain.append(p2i[basename(p)])
 
-    np.savez('train.npz', XTest=XTest, YTest=YTest, XTrain=XTrain, YTrain=YTrain)
+        XTest=np.array(XTest)
+        YTest=np.array(YTest)
+        XTrain=np.array(XTrain)
+        YTrain=np.array(YTrain)
+
+        np.savez('train.npz', XTest=XTest, YTest=YTest, XTrain=XTrain, YTrain=YTrain)
 
     for name, clf in zip(names, classifiers):    
         clf.fit(XTrain, YTrain)
